requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.2
drf-yasg==1.21.7
marshmallow==3.19.0
marshmallow_dataclass==8.5.14
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.28.1
social-auth-app-django==5.2.0
